Move post tab to overview tab, refactor components
